refactor: move GetPrinterList off WebContents
