+++ shell/common/node_bindings.cc
@@ -56,6 +56,7 @@
   V(electron_browser_power_monitor)      \
   V(electron_browser_power_save_blocker) \
   V(electron_browser_protocol)           \
+  V(electron_browser_printing)           \
   V(electron_browser_session)            \
   V(electron_browser_system_preferences) \
   V(electron_browser_base_window)        \
